+++ magick/gem.c
@@ -624,8 +624,8 @@ MagickExport void Hull(const long x_offset,const long y_offset,
       for (x=0; x < (long) columns; x++)
       {
         v=(*p);
-        if (*r <= (v-Upscale(2)))
-          v-=Upscale(1);
+        if (*r <= (v-(long) Upscale(2)))
+          v-=(long) Upscale(1);
         *q=(Quantum) v;
         p++;
         q++;
@@ -661,8 +661,8 @@ MagickExport void Hull(const long x_offset,const long y_offset,
       for (x=0; x < (long) columns; x++)
       {
         v=(*q);
-        if ((*s <= (v-Upscale(2))) && (*r < v))
-          v-=Upscale(1);
+        if ((*s <= (v-(long) Upscale(2))) && (*r < v))
+          v-=(long) Upscale(1);
         *p=(Quantum) v;
         p++;
         q++;
